refactor: Improve Llama.cpp updater and icon
This commit addresses user feedback on the Llama.cpp updater feature:

1.  **Refactor Go Code:** The Go functions and structs related to the updater have been moved from `app.go` to a new, dedicated `llama_updater.go` file. This improves code organization and separation of concerns.

2.  **Update UI Icon:** The rocket emoji (🚀) for the updater button has been replaced with the official Llama.cpp logo. The SVG logo is now included as a local asset and styled to fit with the existing UI buttons.
