+++ local-llm-chat/frontend/index.html
@@ -58,7 +58,7 @@
             <div class="resize-handle"></div> <div class="artifacts-header">
                 <h3>Artifacts</h3>
                 <div class="artifact-actions">
-                    <button id="llamaUpdaterButton" title="Llama.cpp Updater">🚀</button>
+                    <button id="llamaUpdaterButton" title="Llama.cpp Updater"><img src="src/assets/images/llama_logo.svg" alt="Llama.cpp Updater" class="icon" /></button>
                     <button id="mcpManagerButton" title="MCP Manager">🌐</button>
                     <button id="toggleLogViewButton" title="Toggle Log View">📜</button>
                 </div>
