+++ local-llm-chat/frontend/src/style.css
@@ -1252,3 +1252,15 @@ body {
     color: #4caf50;
     font-weight: bold;
 }
+
+#llamaUpdaterButton img.icon {
+    height: 24px;
+    width: auto;
+    vertical-align: middle;
+}
+
+#llamaUpdaterButton {
+    padding: 0;
+    background: none;
+    border: none;
+}
